Refactor: 로그인, 회원가입 화면 클린 아키텍처 적용
