+++ GIL/GIL/Application/SceneDelegate.swift
@@ -8,12 +8,10 @@
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    let appDIContainer = AppDIContainer()
+    let firebaseAuthManager = FirebaseAuthManager()
+    var appFlowCoordinator: AppFlowCoordinator?
     var window: UIWindow?
-    
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
 
     func scene(
         _ scene: UIScene,
@@ -21,24 +19,15 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         options connectionOptions: UIScene.ConnectionOptions
     ) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
-        
-        FirebaseAuthManager.registerAuthStateDidChangeHandler()
-        
-        
-        NotificationCenter.default.addObserver(self, selector: #selector(updateRootViewControllerBasedOnAuthenticationStatus), name: FirebaseAuthManager.authStateDidChangeNotification, object: nil)
-        
-        window = UIWindow(frame: windowScene.coordinateSpace.bounds)
-        window?.backgroundColor = .systemBackground
-        window?.makeKeyAndVisible()
-        window?.windowScene = windowScene
-        
-        updateRootViewControllerBasedOnAuthenticationStatus()
-    }
-    
-    @objc private func updateRootViewControllerBasedOnAuthenticationStatus() {
-        let homeViewController = HomeViewController()
-        let loginViewController = LoginViewController(viewModel: LoginViewModel())
-        window?.rootViewController = UINavigationController(rootViewController: (FirebaseAuthManager.currentUser != nil) ? homeViewController : loginViewController)
+        let window = UIWindow(frame: windowScene.coordinateSpace.bounds)
+        window.backgroundColor = .systemBackground
+        window.windowScene = windowScene
+        self.window = window
+        appFlowCoordinator = AppFlowCoordinator(
+            window: window,
+            appDIContainer: appDIContainer,
+            firebaseAuthManager: firebaseAuthManager
+        )
+        appFlowCoordinator?.start()
     }
 }
-
