--- GIL/GIL/Assets.xcassets/Colors/light_black_dark_white_Color.colorset/Contents.json
@@ -1,28 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "platform" : "universal",
-        "reference" : "labelColor"
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "platform" : "universal",
-        "reference" : "labelColor"
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
