+++ GIL/GIL/Domain/Entities/Member.swift
@@ -1,0 +1,12 @@
+//
+//  Member.swift
+//  GIL
+//
+//  Created by 송우진 on 5/21/24.
+//
+
+import FirebaseAuth
+
+struct Member {
+    var user: User?
+}
